Simplify stringifier registration
